+++ packages/utils/src/str/index.ts
@@ -1,0 +1,14 @@
+import { strEndsWith } from './ends-with';
+import { strStartsWith } from './starts-with';
+
+/**
+ * Utility functions for strings.
+ *
+ * @since 0.0.8
+ */
+export const Str = {
+    startsWith: strStartsWith,
+    endsWith: strEndsWith,
+};
+
+export { strStartsWith, strEndsWith };